separated creating env and initializing database
setup_env.py
